feat(scheduler): optimize edge lookup with FxHashSet for O(1) performance
- Added `edge_set` to `Scheduler` for fast edge existence checks.
- Updated `has_edge_between` to use `edge_set` instead of iterating successors.
- Inserted edges into `edge_set` upon addition to the dependency graph.
- Marked `sub_worlds_dirty` to avoid unnecessary storage cloning.
- Improved performance of `compile()` by reducing edge lookup time.

fix(scripting): update change ticks during component write-back

- Ensured that change ticks are updated when components are modified via Rhai scripts.

fix(serialization): correct row index casting in serializer

- Changed row index casting from `usize` to `usize` in the serializer to prevent potential issues.

feat(optimizations): add detailed optimization plan for APEX ECS

- Documented a comprehensive optimization plan including critical, medium, and long-term improvements.
- Included specific changes, expected effects, and testing requirements for each optimization.

+++ crates/apex-core/src/entity.rs
@@ -19,7 +19,7 @@ impl std::fmt::Display for Entity {
 #[derive(Clone, Copy, Debug)]
 pub struct EntityLocation {
     pub archetype_id: crate::archetype::ArchetypeId,
-    pub row:          usize,
+    pub row:          u32,
 }
 
 struct EntityRecord {
@@ -95,7 +95,7 @@ impl EntityAllocator {
         &mut self,
         entities:     &[Entity],
         archetype_id: crate::archetype::ArchetypeId,
-        start_row:    usize,
+        start_row:    u32,
     ) {
         for (i, entity) in entities.iter().enumerate() {
             let record = &mut self.records[entity.index as usize];
@@ -103,7 +103,7 @@ impl EntityAllocator {
             debug_assert_eq!(record.generation, entity.generation);
             record.location = Some(EntityLocation {
                 archetype_id,
-                row: start_row + i,
+                row: start_row + i as u32,
             });
         }
     }
@@ -238,7 +238,7 @@ mod tests {
         for (i, entity) in entities.iter().enumerate() {
             let loc = alloc.get_location(*entity).unwrap();
             assert_eq!(loc.archetype_id.0, 42);
-            assert_eq!(loc.row, i);
+            assert_eq!(loc.row as usize, i);
         }
     }
 }
